# PERF: Improve to-markdown speed, update the code